pyproject.toml
[tool.poetry]
name = "vjemmie"
version = "0.1.0"
description = ""
readme = "README.md"
authors = ["PederHA <peder.andresen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
praw = "^6.5.1"
numpy = "^1.18.1"
gtts = "^2.1.1"
xmltodict = "^0.12.0"
geopy = "^1.21.0"
pynacl = "^1.3.0"
trueskill = "^0.4.5"
pathvalidate = "^2.2.0"
recordclass = "^0.13.2"
urllib3 = "^1.25.8"
pygithub = "^1.46"
requests-html = "^0.10.0"
markovify = "^0.8.0"
psutil = "^5.6.7"
beautifulsoup4 = "^4.8.2"
ciso8601 = "^2.1.3"
websockets = "^8.1"
spotipy = "^2.9.0"
google-api-python-client = "^1.7.11"
pytesseract = "^0.3.2"
python-magic = "^0.4.15"
unidecode = "^1.1.1"
mcstatus = "^2.3.0"
mwthesaurus = "^0.1.2"
httpx = "^0.13.3"
aiofile = "^3.1.0"
pytz = "^2020.4"
Pillow = "^8.0.1"
"discord.py" = "^1.5.1"
youtube-dl = "2021.6.6"
aiomysql = "^0.0.21"
pydantic = "^1.8.2"
uvloop = { version="^0.15.2", platform="linux" }
vjemmieapi = "^0.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"
mypy = "^0.780"
autopep8 = "^1.5.3"
ptvsd = "^4.3.2"
black = "^21.6b0"
pytest-asyncio = "^0.15.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
